@@ -1,24 +1,24 @@
 {
-    "name": "sparkpost/php-sparkpost",
-    "description": "SDK for interfacing with SparkPost APIs",
-    "license": "Apache 2.0",
-    "authors": [
-        {
-            "name": "Message Systems, Inc."
-        }
-    ],
-    "minimum-stability": "stable",
-    "require": {
-    	"php": ">=5.3.0",
-        "guzzlehttp/guzzle": "5.3.*"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "4.3.*"
-    },
-    "autoload": {
-    	"psr-4": {
-    		"SparkPost\\": "lib/SparkPost/",
-    		"SparkPost\\SendGridCompatibility\\": "lib/SendGridCompatibility/"
-    	}
+  "name": "sparkpost/php-sparkpost",
+  "description": "SDK for interfacing with SparkPost APIs",
+  "license": "Apache 2.0",
+  "authors": [
+    {
+      "name": "Message Systems, Inc."
     }
+  ],
+  "minimum-stability": "stable",
+  "require": {
+  	"php": ">=5.5.0",
+    "guzzlehttp/guzzle": "6.0.*"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "4.8.*"
+  },
+  "autoload": {
+  	"psr-4": {
+  		"SparkPost\\": "lib/SparkPost/",
+  		"SparkPost\\SendGridCompatibility\\": "lib/SendGridCompatibility/"
+  	}
+  }
 }
