@@ -2,29 +2,29 @@
 namespace SparkPost;
 
 class SparkPost {
-	
+
 	private static $config;
-	private static $defaults = array(
+	private static $defaults = [
 			'host'=>'api.sparkpost.com',
 			'protocol'=>'https',
 			'port'=>443,
 			'strictSSL'=>true,
 			'key'=>'',
-			'version'=>'v1'	
-		);
-	
+			'version'=>'v1'
+		];
+
 	/**
 	 * Enforce that this object can't be instansiated
 	 */
 	private function __construct(){}
-	
+
 	/**
 	 * Allows the user to pass in values to override the defaults and set their API key
 	 * @param Array $configMap - Hashmap that contains config values for the SDK to connect to SparkPost
 	 * @throws \Exception
 	 */
 	public static function setConfig(array $configMap) {
-		//check for API key because its required	
+		//check for API key because its required
 		if (isset($configMap['key'])){
 			$key = trim($configMap['key']);
 			if(empty($key)){
@@ -40,13 +40,13 @@ class SparkPost {
 			}
 		}
 	}
-	
+
 	/**
 	 * Retrieves the configuration that was previously setup by the user
 	 * @throws \Exception
 	 */
 	public static function getConfig() {
-		if (self::$config === null) {	
+		if (self::$config === null) {
 			throw new \Exception('No configuration has been provided');
 		}
 		return self::$config;
