@@ -4,8 +4,8 @@ require_once (dirname(__FILE__).'/../bootstrap.php');
 use SparkPost\SparkPost;
 use SparkPost\Transmission;
 
-$key = 'YOURAPIKEY'; 
-SparkPost::setConfig(array('key'=>$key));
+$key = 'YOURAPIKEY';
+SparkPost::setConfig(['key'=>$key]);
 
 try {
 	$results = Transmission::find('Your Transmission Id');
