@@ -5,37 +5,37 @@ use SparkPost\SparkPost;
 use SparkPost\Transmission;
 
 $key = 'YOURAPIKEY';
-SparkPost::setConfig(array('key'=>$key));
+SparkPost::setConfig(['key'=>$key]);
 
 try{
-	$results = Transmission::send(array(
+	$results = Transmission::send([
 	  "campaign"=>"my-campaign",
-	  "metadata"=>array(
+	  "metadata"=>[
 	   "sample_campaign"=>true,
 	   "type"=>"these are custom fields"
-	  ),
-	  "substitutionData"=>array(
+	 	],
+	  "substitutionData"=>[
 	    "name"=>"Test Name"
-	  ),
+	  ],
 	  "description"=>"my description",
 	  "replyTo"=>"reply@test.com",
-	  "customHeaders"=>array(
+	  "customHeaders"=>[
 	    "X-Custom-Header"=>"Sample Custom Header"
-	  ),
+	  ],
 	  "trackOpens"=>false,
 	  "trackClicks"=>false,
-	  "from"=>"From Envelope <from@example.com>",
+	  "from"=>"From Envelope <sandbox@sparkpostbox.com>",
 	  "html"=>"<p>Hello World! Your name is: {{name}}</p>",
 	  "text"=>"Hello World!",
 	  "subject"=>"Example Email: {{name}}",
-	  "recipients"=>array(
-	    array(
-	      "address"=>array(
+	  "recipients"=>[
+	  	[
+	    	"address"=>[
 	        "email"=>"john.doe@sample.com"
-	      )
-	    )
-	  )
-	));
+	      ]
+	    ]
+	  ]
+	]);
 	echo 'Congrats you can use your SDK!';
 } catch (\Exception $exception) {
 	echo $exception->getMessage();
