@@ -4,21 +4,21 @@ require_once (dirname(__FILE__).'/../bootstrap.php');
 use SparkPost\SparkPost;
 use SparkPost\Transmission;
 
-$key = 'YOURAPIKEY'; 
-SparkPost::setConfig(array('key'=>$key));
+$key = 'YOURAPIKEY';
+SparkPost::setConfig(['key'=>$key]);
 
 try {
-	$results = Transmission::send(array(
-		"from"=>"From Envelope <from@example.com>",
-		"recipients"=>array(
-			array(
-			"address"=>array(
+	$results = Transmission::send([
+		"from"=>"From Envelope <sandbox@sparkpostbox.com>",
+		"recipients"=>[
+			[
+				"address"=>[
 					"email"=>"john.doe@sample.com"
-				)
-			)
-		),
+				]
+			]
+		],
 		"template"=>"my-template"
-	));
+	]);
 	echo 'Congrats you can use your SDK!';
 } catch (\Exception $exception) {
 	echo $exception->getMessage();
