@@ -4,15 +4,15 @@ namespace SparkPost\Test;
 use SparkPost\SparkPost;
 
 class SparkPostTest extends \PHPUnit_Framework_TestCase {
-	
+
 	/**
 	 * @desc Ensures that the configuration class is not instantiable.
 	 */
 	public function testConstructorCannotBeCalled() {
 		$class = new \ReflectionClass('\SparkPost\SparkPost');
-		$this->assertFalse($class->isInstantiable()); 
+		$this->assertFalse($class->isInstantiable());
 	}
-	
+
 	/**
 	 * @desc Tests that an exception is thrown when a library tries to recieve the config and it has not yet been set.
 	 * 		Since its a singleton this test must come before any setConfig tests.
@@ -22,31 +22,31 @@ class SparkPostTest extends \PHPUnit_Framework_TestCase {
 	public function testGetConfigEmptyException() {
 		SparkPost::getConfig();
 	}
-	
+
 	/**
 	 * @desc Tests that the api key is set when setting the config
 	 * @expectedException Exception
 	 * @expectedExceptionMessage You must provide an API key
 	 */
 	public function testSetConfigAPIKeyNotSetException() {
-		SparkPost::setConfig(array('something'=>'other than an API Key'));
+		SparkPost::setConfig(['something'=>'other than an API Key']);
 	}
-	
+
 	/**
 	 * @desc Tests that the api key is set when setting the config and that its not empty
 	 * @expectedException Exception
 	 * @expectedExceptionMessage You must provide an API key
 	 */
 	public function testSetConfigAPIKeyEmptyException() {
-		SparkPost::setConfig(array('key'=>''));
+		SparkPost::setConfig(['key'=>'']);
 	}
-	
+
 	/**
 	 * @desc Tests overridable values are set while invalid values are ignored
 	 */
 	public function testSetConfigMultipleValuesAndGetConfig() {
-		SparkPost::setConfig(array('key'=>'lala', 'version'=>'v8', 'port'=>1024, 'someOtherValue'=>'fakeValue'));
-		
+		SparkPost::setConfig(['key'=>'lala', 'version'=>'v8', 'port'=>1024, 'someOtherValue'=>'fakeValue']);
+
 		$testConfig = SparkPost::getConfig();
 		$this->assertEquals('lala', $testConfig['key']);
 		$this->assertEquals('v8', $testConfig['version']);
